changes to help text and other minor things

+++ musicdownloader.php
@@ -3,33 +3,50 @@ include_once 'src/downloader.php';
 include_once 'src/logger.php';
 include_once 'src/cacher.php';
 include_once 'src/utils.php';
-$helpText = "Help \n -init <name>: creates a folder with the inserted name, if name is left empty, the folder name will be downloads; \n 
--config <> \n";
-$howToUseText = "how to use";
 
-
-function parseArguments($arguments){
-    
+function parseArguments($arguments)
+{
     $mdLogger=new Logger("Music Downloader");
     $isFlag = substr($arguments[1], 0 ,1) == "-";
-    if (!$isFlag){ 
+    if (!$isFlag)
+    { 
         download($arguments, $mdLogger);
     } else {
         $mdLogger->info("Called the \"".$arguments[1]."\" flag");
-        switch ($arguments[1]) {
-            case "-init": 
-                
-                break;
+        switch ($arguments[1]) 
+        {
             case "-help": 
-                global $helpText;
-                echo $helpText;
+                echo(PHP_EOL."The following text is supposed to help you use this script correctly: ".PHP_EOL."\n\t".
+                "To download a video, simply paste the link after calling the script \n\t".
+                "You can add flags(-mp4, webm, 3gp) to identify which extension you prefer \n\t".
+                "If no flag is given, the script will automatically download the link with the best quality \n\t".
+                "Example 1: php musicdownloader.php \"https://www.youtube.com/watch?v=tohW1VLqCFo\" \n\t".
+                "Example 2: php musicdownloader.php \"https://www.youtube.com/watch?v=tohW1VLqCFo\" -mp4".PHP_EOL."\n\t".
+                "The script can also be used to download entire playlists from youtube, like downloading \n\t".
+                "individual videos, you can use flags to specify which extension is preferred \n\t".
+                "Example 1: php musicdownloader.php \"https://www.youtube.com/playlist?list=PLVrHtnQEDQVbrm-d1PmFqS-zw4FTT_7hp\" \n\t".
+                "Example 2: php musicdownloader.php \"https://www.youtube.com/playlist?list=PLVrHtnQEDQVbrm-d1PmFqS-zw4FTT_7hp\" -mp4".PHP_EOL."\n\t".
+                "You can also configure the script to download to a specific folder or to always download using a specific flag \n\t".
+                "Example 1: Specify download directory -> php musicdownloader.php -config -dir \"-insert-dir-name-here-\" \n\t".
+                "Example 2: Specify default flags -> php musicdownloader.php -config -flags \"-insert-flag-here-\" \n\t".
+                "Example 3: Clear all configurations -> php musicdownloader.php -config -clear".PHP_EOL."\n\t".
+                "Possible flags used for downloading videos: \n\t".
+                "\t-v: Video only; \n\t".
+                "\t-a: Audio only; \n\t".
+                "\t-av: Audio and Video; \n\t".
+                "\t-mp4: Extension Mp4; \n\t".
+                "\t-3gp: Extension 3gp; \n\t".
+                "\t-webm: Extension webm; \n\t".
+                "\t-m4a: Extension m4a; \n\t");
                 break;
             case "-config": 
                 config($arguments, $mdLogger);
                 break;
+            case "-clear-logs": 
+                $mdLogger->clear();
+                break;
             default: 
-                global $howToUseText;
-                echo $howToUseText;
+                echo("Invalid call to script, please use \"php musicdownloader.php -help\" for general information");
                 break;
         }
     }
@@ -62,8 +79,7 @@ function config($mArg, $dLogger){
             case "-clear":
                 $files = glob($cacher->fetch("directory").'/*'); // get all file names
                 foreach($files as $file){ // iterate files
-                    if(is_file($file))
-                    unlink($file); // delete file
+                    if(is_file($file)) unlink($file); // delete file
                 }
             //podemos incluir tbm
             //$cacher->store("directory", "");*/
@@ -72,16 +88,6 @@ function config($mArg, $dLogger){
             break;
         }
         $dLogger->info("Config argument \"".$mArg[2]."\"");
-}
-/*
-    **  Function to create a folder to hold the downloads
-    **  @param $name -> name of the folder
-    **  @param $dLogger -> instance of logger
-    **  @throws Exception -> If the folder couldn't be created
-    **  @return -> none
-    */
-function init($name, $dLogger){
-    
 }
 
 function download($arguments, $dLogger){
